requirements.txt
flask
pyyaml
qbittorrent-api
tmdbv3api
tvdbsimple
uwsgi
